### 2025-04-24 #8104 @editorialbot
[PRE REVIEW]: Bamojax: Bayesian Modelling with JAX
**Submitting author:** @maxhinne (<a href="http://orcid.org/0000-0002-9279-6725">Max Hinne</a>)
**Repository:** https://github.com/UncertaintyInComplexSystems/bamojax
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.1.0
