### 2025-04-24 #8104 @crvernon
@maxhinne - Please check out the paper compile error related to the ORCID mentioned above in the thread:  https://github.com/openjournals/joss-reviews/issues/8104#issuecomment-2828058307
