+++ paper/paper.bib
@@ -22,7 +22,7 @@
 @article{Hinne2025,
 	title = {An introduction to {Sequential} {Monte} {Carlo} for {Bayesian} inference and model comparison—with examples for psychology and behavioral science},
 	volume = {57},
-	doi = {https://doi.org/10.3758/s13428-025-02642-1},
+	doi = {10.3758/s13428-025-02642-1},
 	abstract = {Bayesian inference is becoming an increasingly popular framework for statistics in the behavioral sciences. However, its application is hampered by its computational intractability – almost all Bayesian analyses require a form of approximation. While some of these approximate inference algorithms, such as Markov chain Monte Carlo (MCMC), have become well known throughout the literature, other approaches exist that are not as widespread. Here, we provide an introduction to another family of approximate inference techniques known as Sequential Monte Carlo (SMC). We show that SMC brings a number of beneﬁts, which we illustrate in three different examples: linear regression and variable selection for depression, growth curve mixture modeling of grade point averages, and in computational modeling of the Iowa Gambling Task. These use cases demonstrate that SMC is efﬁcient in exploring posterior distributions, reaching similar predictive performance as state-of-the-art MCMC approaches in less wall-clock time. Moreover, they show that SMC is effective in dealing with multi-modal distributions, and that SMC not only approximates the posterior distribution but simultaneously provides a useful estimate of the marginal likelihood, which is the essential quantity in Bayesian model comparison. All of this comes at no additional effort from the end user.},
 	language = {en},
 	number = {25},
@@ -30,9 +30,13 @@
 	author = {Hinne, Max},
 	year = {2025}
 }
-@software{pymc,
+@article{pymc,
   title = {{PyMC}: A Modern and Comprehensive Probabilistic Programming Framework in {P}ython},
   author = {Abril-Pla O, Andreani V, Carroll C, Dong L, Fonnesbeck CJ, Kochurov M, Kumar R, Lao J, Luhmann CC, Martin OA, Osthege M, Vieira R, Wiecki T, Zinkov R.},
+  journal = {{PeerJ} Computer Science},
+  volume = {9},
+  number = {e1516},
+  doi = {10.7717/peerj-cs.1516},
   year = {2023},
   url = {https://www.pymc.io/}
 }
