### 2025-04-24 #8104 @mhinne
@crvernon Thank you, I've removed the https://orcid..org/ prefix, so I think it should compile now.

I also fixed the invalid DOI.

As for potential reviewers, I see several with Bayesian statistics expertise, such as Virgile Andreani, Daniel Waxman, and Olga Doronina. Erik-Jan van Kesteren would also be very appropriate, but is somewhat close (we have shared co-authors).
