### 2025-06-24 c1c03bf
Add reversible-jump MCMC; bump to v0.2.0
