+++ bamojax/marginal_likelihoods.py
@@ -1,0 +1,215 @@
+
+import jax
+import jax.numpy as jnp
+import jax.random as jrnd
+
+import jaxopt
+from tqdm import tqdm
+
+from jax.scipy.special import logsumexp
+from jax.tree_util import tree_flatten, tree_unflatten, tree_map
+from distrax._src.distributions.distribution import Distribution
+from distrax._src.bijectors.bijector import Bijector
+
+from jaxtyping import Float
+from typing import Callable, Dict
+
+from bamojax.base import Model
+
+# TODO: implement testing for these functions
+
+def iid_likelihood(L: Callable):
+    r"""
+    
+    We typically have multiple observations and assume the likelihood factorizes 
+    as: 
+    
+        \log p(Y \mid \theta) = \sum_{i=1}^N \log p(y_i \mid \theta) .
+
+    """
+    return lambda x: jnp.sum(L()(x))
+
+#
+def naive_monte_carlo(key, 
+                      model: Model, 
+                      num_prior_draws: int = 1_000, 
+                      num_chunks: int = 5,
+                      iid_obs: bool = True,
+                      pb = True) -> Float:   
+    r"""The Naive Monte Carlo (NMC) estimator
+
+    The marginal likelihood is defined by 
+    
+        p(D) = \int_\Theta p(D \mid \theta) p(\theta) d\theta.
+
+    In NMC we draw samples from the prior and approximate the ML as
+
+        p(D) \approx 1/N \sum_{i=1}^N p(D \mid \theta_i), with \theta_i ~ p(\theta).
+
+    In nontrivial models, we need a *large* N for this approximation to be 
+    reasonable.
+
+    """
+
+    if iid_obs:
+        loglikelihood_fn = iid_likelihood(model.loglikelihood_fn)
+    else:
+        loglikelihood_fn = model.loglikelihood_fn
+
+    loglikelihoods = jnp.zeros((num_prior_draws, 
+                                num_chunks))
+    
+    # We don't want to vmap this loop, as the reason for the loop is to avoid
+    # running out of memory!
+    for i in tqdm(range(num_chunks), disable=not pb):
+        key, subkey = jrnd.split(key)
+        keys = jrnd.split(subkey, num_prior_draws)
+        prior_draws = jax.vmap(model.sample_prior)(keys)
+        loglikelihoods = loglikelihoods.at[:, i].set(jax.vmap(loglikelihood_fn)(prior_draws))
+    return logsumexp(loglikelihoods.flatten()) - jnp.log(num_prior_draws*num_chunks)
+
+#
+def harmonic_mean(key, model: Model) -> Float:
+    raise NotImplementedError
+
+def generalized_harmonic_mean(key, model: Model) -> Float:
+    raise NotImplementedError
+
+#
+def importance_sampling(key, 
+                        model: Model, 
+                        g_IS: Distribution,
+                        num_samples: int = 1_000,
+                        iid_obs: bool = True) -> Float:
+    
+    r"""Importance sampling routine for a given BayesianModel.
+
+    Importance sampling is based around the following approximation to the log
+    marginal likelihood (see e.g., Gronau et al., 2017):
+
+    p(D) \approx 1/N \sum_{i=1}^N p(D \mid \theta_i) p(\theta_i) / g_IS(\theta_i),
+
+    with
+
+    \theta_i ~ g_IS(\theta)
+
+    Here, g_IS is the importance density, which should meet these criteria:
+
+    1. It is easy to evaluate.
+    2. It has the same domain as the posterior p(\theta \mid D).
+    3. It matches the posterior as closely as possible.
+    4. It has fatter tails than the posterior.
+
[…]
