+++ bamojax/samplers.py
@@ -1,10 +1,12 @@
-from typing import NamedTuple
+from typing import NamedTuple, Callable
 import jax
 import jax.random as jrnd
 import jax.numpy as jnp
 from blackjax.base import SamplingAlgorithm
 from blackjax.types import ArrayTree, PRNGKey
 from blackjax import normal_random_walk, generate_top_level_api_from
+from distrax import Distribution
+
 from .base import Model
 from .modified_blackjax import modified_elliptical_slice_nd
 
@@ -17,6 +19,11 @@ class GibbsState(NamedTuple):
 
 #
 class MCMCState(NamedTuple):
+
+    position: ArrayTree
+
+#
+class RJState(NamedTuple):
 
     position: ArrayTree
 
@@ -227,6 +234,184 @@ def mcmc_sampler(model: Model,
         return state, info
     
     #
+    return SamplingAlgorithm(init_fn, step_fn)
+
+#
+def reversible_jump_mcmc(models: list[Model], 
+                         auxiliary_proposal_dist: Distribution,
+                         jump_functions: list[Callable],
+                         jacobians: list[Callable],
+                         projections: list[Callable],
+                         within_model_kernels: list[dict],
+                         model_move_prob: float = 0.5) -> SamplingAlgorithm:
+    """Constructs a reversible jump MCMC algorithm for the given models.
+    
+    Implementation follows the reversible jump MCMC algorithm described by Hastie & Green (2012).
+
+    Args:
+        models: List of models to sample from.
+        auxiliary_proposal_dist: Distribution to sample the auxiliary variable from.
+        jump_functions: List of functions to transform the position when jumping between models.
+        jacobians: List of Jacobian determinant functions for the jump functions
+        within_model_kernel: List of within-model sampling algorithms, one for each model.
+
+    Returns:
+        SamplingAlgorithm: A sampling algorithm that implements the reversible jump MCMC.
+
+    References:
+        Hastie, T., & Green, P. J. (2012). Model choice using reversible jump Markov chain Monte Carlo
+    
+    """
+
+    assert len(models) == 2, 'Reversible jump MCMC currently only supports two models.'
+    assert len(jump_functions) == 2, 'Reversible jump MCMC requires two jump functions for the two models.'
+    assert len(jacobians) == 2, 'Reversible jump MCMC requires two Jacobian determinant functions for the two models.'
+    assert len(within_model_kernels) == 2, 'Reversible jump MCMC requires within-model sampling algorithms for each models.'
+
+    def make_within_model_kernel(model_index) -> SamplingAlgorithm:
+        """Creates a within-model kernel for the specified model index.
+
+        Args:
+            model_index: Index of the model for which to create the within-model kernel.
+
+        Returns:
+            A SamplingAlgorithm that performs within-model sampling for the specified model.
+
+        """
+        return mcmc_sampler(models[model_index], 
+                            mcmc_kernel=within_model_kernels[model_index]['mcmc_kernel'], 
+                            mcmc_parameters=within_model_kernels[model_index]['mcmc_parameters'])
+    
+    #    
+    def make_within_model_fn(model_index) -> tuple[RJState, dict]:
+        """Creates a function that performs a within-model move for the specified model index.
+
+        Args:
+            model_index: Index of the model for which to create the within-model move function.
+
+        Returns:
+            A function that takes a key, position, and optional arguments, and returns a new RJState and info dictionary.
+            The info dictionary imputes nan values for log probabilities and Jacobian determinants, as they are not used in within-model moves.
+
+        """
+        def fn( key, position, *args):
+            temperature = args[0] if len(args) > 0 else 1.0
+            kernel = mcmc_samplers[model_index]
[…]

[…]
