### 2025-06-25 e1ea668
Add Laplace approximation to linear regression example; fix Laplace bijectors
