+++ bamojax/inference.py
@@ -24,6 +24,10 @@ import jax.numpy as jnp
 from jax.tree_util import tree_map, tree_flatten, tree_unflatten
 from jax.flatten_util import ravel_pytree
 
+from tensorflow_probability.substrates import jax as tfp
+tfd = tfp.distributions
+tfb = tfp.bijectors
+
 def run_window_adaptation(model, key: PRNGKey, num_warmup_steps):
     r""" Find optimized parameters for HMC-based inference.
 
@@ -47,6 +51,45 @@ def run_window_adaptation(model, key: PRNGKey, num_warmup_steps):
     return warm_state, warm_parameters
 
 #
+def get_model_bijectors(model) -> dict:
+    r""" Meanfield VI imposes a Gaussian variational distribution. 
+    
+    In order to use the correct parameter constraints this function determines all relevant bijectors.
+
+    It's a bijector detector! :-)
+
+    Returns:
+        A dictionary with bijectors for the variables that use it, and an identity bijector otherwise.
+    
+    """
+    latent_nodes = model.get_latent_nodes()
+
+    # forward_transforms = {}
+    # backward_transforms = {}
+    # for node_name, node in latent_nodes.items():
+    #     if hasattr(node.distribution, '_bijector'):
+    #         bij = node.distribution._bijector
+    #         forward_transform = lambda x, bij=bij: bij.forward(x)
+    #         backward_transform = lambda x, bij=bij: bij.inverse(x)
+    #     else:
+    #         forward_transform = lambda x: x
+    #         backward_transform = lambda x: x
+    #     forward_transforms[node_name] = forward_transform
+    #     backward_transforms[node_name] = backward_transform
+    # return forward_transforms, backward_transforms
+    bijectors = {}
+    for node_name, node in latent_nodes.items():
+        if hasattr(node.distribution, '_bijector'):
+            bij = node.distribution._bijector
+            transform = bij
+        else:
+            # transform = dx.Independent(tfb.Identity(), reinterpreted_batch_ndims=1)
+            transform = tfb.Identity()  # Use TensorFlow Probability's Identity bijector
+            # transform = dx.Lambda(forward = lambda x: x, inverse = lambda x: x)  # Use Distrax's Lambda bijector
+        bijectors[node_name] = transform
+    return bijectors
+
+# 
 
 class InferenceEngine(ABC):
     r""" The abstract class for inference engines
@@ -496,39 +539,19 @@ class VIInference(InferenceEngine):
             if not isinstance(optimizer_chain_args, list):
                 optimizer_chain_args = [optimizer_chain_args]
             self.optimizer = optax.chain(*optimizer_chain_args, optimizer)
-        self.bijectors = self.get_model_bijectors()
+        self.bijectors = get_model_bijectors(self.model)
+        self.is_leaf_fn = lambda x: hasattr(x, 'forward') and hasattr(x, 'inverse')
+        self.forward_bijectors = lambda x: jax.tree.map(lambda b, v: b.forward(v), self.bijectors, x, is_leaf=self.is_leaf_fn)
+        self.backward_bijectors = lambda x: jax.tree.map(lambda b, v: b.inverse(v), self.bijectors, x, is_leaf=self.is_leaf_fn)
+
         def logdensity_fn(z):
-            z = jax.tree.map(lambda f, v: f(v), self.bijectors, z)
+            z = self.forward_bijectors(z)
             return model.loglikelihood_fn()(z) + model.logprior_fn()(z)
 
         #
         self.logdensity_fn = logdensity_fn
 
-    #
-    def get_model_bijectors(self) -> dict:
-        r""" Meanfield VI imposes a Gaussian variational distribution. 
-        
[…]
