### 2025-08-12 UncertaintyInComplexSystems/bamojax#2 @matt-graham
Publishing API documentation
Raising as part of openjournals/joss-reviews#8642

It looks like while there are docstrings in the classes / functions defined in the package these are not currently visible without browsing the source code directly. It would be helpful to use something like [Sphinx](https://www.sphinx-doc.org/en/master/index.html) (with [`autodoc` extension](https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html#module-sphinx.ext.autodoc)) or [MkDocs](https://www.mkdocs.org/) (with [`mkdocstrings-python` plugin](https://mkdocstrings.github.io/python/)) to automatically generate a documentation website to make it easier for users to browse the documentation - this could be hosted on [GitHub Pages](https://docs.github.com/en/pages) or a third-party service like [ReadTheDocs](https://about.readthedocs.com/).
