### 2025-08-12 UncertaintyInComplexSystems/bamojax#3 @matt-graham
Instructions for running tests locally
Raising as part of openjournals/joss-reviews#8642

It would be good to include some brief instructions for how developers / users should run the test suite, including dependencies they need to install. While this can be worked out from the GitHub Actions test workflow file, making it easier for people to run tests means its more likely they will actually check they pass!

Relatedly, to simplify setting up development environment with dependencies required for running tests, it might make sense to define `dev` [optional dependencies](https://setuptools.pypa.io/en/stable/userguide/dependency_management.html#optional-dependencies) or [dependency group](https://packaging.python.org/en/latest/specifications/dependency-groups/) (might require transitioning from `setup.py` / `setup.cfg` to using a `pyproject.toml` file) to your `setuptools` configuration with the additional test depenedencies specified (I think `pytest` and `git+https://github.com/UncertaintyInComplexSystems/JaxKern.git`).

This would then allow doing something like (if using optional dependencies)

```
pip install -e ".[dev]"
```

or (if using dependency group)

```
pip install --group dev -e .
```

to install both package (in editable mode) and test dependencies, and mean this can be included in both the GitHub Actions workflow file and instructions for running tests without needing to duplicate specification of test dependencies.
