### 2025-08-12 UncertaintyInComplexSystems/bamojax#4 @matt-graham
Community guidelines
Raising as part of openjournals/joss-reviews#8642

Currently there are no guidelines (that I could spot at least!) for external parties wanting to contribute to the package, how to report issues or problems or seek support.
