### 2025-08-12 d615d0b
Fix typos and expand acronyms in README
Part of openjournals/joss-reviews#8642

Suggests a few minor edits to README to correct a couple of typos I noticed and define / expand some acronyms on first usage.
