### 2025-08-12 UncertaintyInComplexSystems/bamojax#6 @matt-graham
Quick start example in README does not run
Raising as part of openjournals/joss-reviews#8642

The quick start example in the README has some missing imports and at least one syntax error.

- [ ] Missing `import distrax as dx`
- [ ] Missing `import blackjax`
- [ ] Missing `from bamojax.samplers import mcmc_sampler`
- [ ] Syntax error in line
     ```Python
     observations = my_model.add_node('x', distribution=dx.Bernoulli, parents=dict(probs=latent_theta), link_fn, observations=x)
     ```
     due to positional argument `link_fn` following keyword argument.
- [ ] Not clear where name `smc_inference_loop` is defined as can't find any instances of it in package code beyond one of the example notebooks.
