### 2025-08-12 UncertaintyInComplexSystems/bamojax#7 @matt-graham
`__init__.py` in root of repository should not be needed
Raising as part of openjournals/joss-reviews#8642

It is not clear why there an empty `__init__.py` file in the root of the repository. In general the presence of an `__init__.py` file in a directory makes Python consider the directory to be a package. Here the `bamojax` subdirectory contains a `__init__.py` file which makes it recognised as a package, but the empty `__init__.py` file in the directory above shouldn't be necessary and may cause issues with imports if run a Python interpreter at the root of the repository or directory above (as the top-level directory containing the package would be considered a package on the local path and so might block imports to the actual `bamojax` package which would then be considered a subpackage).
