### 2025-08-19 UncertaintyInComplexSystems/bamojax#8 @kanishkan91
JOSS Review: Mention python version required explicitly in README
Since this package cannot be installed via pip, its a bit unclear which python version to use. While this is documented in the setup.py (>=3.10), this should also be mentioned in the README for user convenience.

https://github.com/openjournals/joss-reviews/issues/8642
