### 2025-09-14 UncertaintyInComplexSystems/bamojax#6 @mhinne
Thank you for pointing this out; that is a rather silly mistake. I have now updated the README to include the correct imports and use the up-to-date way of calling inference in bamojax.
