### 2025-09-14 4629183
Update README.md
Updates quick start example (fixes #6)
