### 2025-09-14 0f3b153
Delete __init__.py
Remove redundant (and potentially troublesome) __init__.py in root (fixes #7)
