### 2025-09-14 UncertaintyInComplexSystems/bamojax#8 @mhinne
Thank you, that is a good point. The README now refers to the required Python version (>= 3.10) in the installation instructions.
