### 2025-09-14 UncertaintyInComplexSystems/bamojax#1 @mhinne
Thank you for the helpful comments! The textual mistakes I have updated right away.

I agree that the statements on comparing bamojax to the mentioned tools may have been a bit overzealous. I have rewritten the comparison section somewhat to better reflect the aim of bamojax (allowing for all the fast Blackjax algorithms but also stay close to JAX Numpy for method development).

Please let me know if you agree with these changes.
