### 2025-09-14 ede399a
Updated JOSS paper
Updated the paper to fix minor mistakes and better reflect the mission statement of bamojax (refs #1 )
