+++ paper/paper.bib
@@ -30,6 +30,34 @@
 	author = {Hinne, Max},
 	year = {2025}
 }
+@article{Hoffman2014,
+  author  = {Matthew D. Hoffman and Andrew Gelman},
+  title   = {{The No-U-Turn Sampler}: Adaptively Setting Path Lengths in {Hamiltonian Monte Carlo}},
+  journal = {Journal of Machine Learning Research},
+  year    = {2014},
+  volume  = {15},
+  number  = {47},
+  pages   = {1593--1623},
+  url     = {http://jmlr.org/papers/v15/hoffman14a.html}
+}
+
+@inproceedings{Murray2010,
+  title = 	 {Elliptical slice sampling},
+  author = 	 {Murray, Iain and Adams, Ryan and MacKay, David},
+  booktitle = 	 {Proceedings of the Thirteenth International Conference on Artificial Intelligence and Statistics},
+  pages = 	 {541--548},
+  year = 	 {2010},
+  editor = 	 {Teh, Yee Whye and Titterington, Mike},
+  volume = 	 {9},
+  series = 	 {Proceedings of Machine Learning Research},
+  address = 	 {Chia Laguna Resort, Sardinia, Italy},
+  month = 	 {13--15 May},
+  publisher =    {PMLR},
+  pdf = 	 {http://proceedings.mlr.press/v9/murray10a/murray10a.pdf},
+  url = 	 {https://proceedings.mlr.press/v9/murray10a.html}
+}
+
+
 @article{pymc,
   title = {{PyMC}: A Modern and Comprehensive Probabilistic Programming Framework in {P}ython},
   author = {Abril-Pla O, Andreani V, Carroll C, Dong L, Fonnesbeck CJ, Kochurov M, Kumar R, Lao J, Luhmann CC, Martin OA, Osthege M, Vieira R, Wiecki T, Zinkov R.},
