+++ paper/paper.md
@@ -27,7 +27,7 @@ penalties, and the ability to quantify evidence in favour of the null hypothesis
 
 However, exact Bayesian inference is computationally intractable in all but the
 simplest of cases, and requires _approximate inference_ techniques, such as 
-Markov chain Monte Carlo and variational inference. Recent advances in the Python
+Markov chain Monte Carlo (MCMC) and variational inference. Recent advances in the Python
 JAX [@jax] framework have enabled highly efficient implementations of these algorithms,
 due to features such as automated differentation and GPU acceleration. These 
 developments have the potential to greatly increase the efficiency of statistical 
@@ -41,14 +41,14 @@ modelling pipelines.
 **Bamojax** is a Bayesian modelling tool based on Python & JAX [@jax]. It
 provides an intuitive, intermediate-level interface between defining a Bayesian 
 statistical model conceptually, and performing efficient inference using the 
-Blackjax package [@blackjax]. 
+BlackJAX package [@blackjax]. 
 
-Existing probabilistic programming languages, such as PyMC [@pymc], can export a logdensity function that enables Blackjax-based inference. However, this has two limitations:
+Existing probabilistic programming languages, such as PyMC [@pymc], can export (the logarithm of) a probability density function that enables BlackJAX-based inference. However, this has two limitations:
 
-1. It does not support Gibbs sampling, where variables are updated individually using their own MCMC kernels. For example, when approximating the posterior over a latent Gaussian process and its hyperparameters, elliptical slice sampling for the GP is often more efficient than applying NUTS to all variables jointly. This becomes even more important when embedding MCMC sampling in Sequential Monte Carlo [@Hinne2025].
-2. It makes it harder to apply tempered Sequential Monte Carlo methods that need separate prior and likelihood densities.
+1. It does not support Gibbs sampling, where variables are updated individually using their own MCMC kernels. For example, when approximating the posterior over a latent Gaussian process (GP) and its hyperparameters, elliptical slice sampling [@Murray2010] for the GP can be more efficient than applying No-U-Turn Hamiltonian Monte Carlo (NUTS HMC; [@Hoffman2014]) sampling  to all variables jointly. This becomes even more important when embedding MCMC sampling in Sequential Monte Carlo [@Hinne2025].
+2. It makes it harder to apply tempered Sequential Monte Carlo methods that need separate prior and likelihood terms.
 
-While users can circumvent these issues by manually implementing their models using Blackjax,  this is a labor-intensive and error-prone process. **Bamojax** addresses this gap by providing a user-friendly interface for model construction and Gibbs sampling on top of Blackjax.
+While users can circumvent these issues by manually implementing their models using BlackJAX,  this is a labor-intensive and error-prone process. **Bamojax** addresses this gap by providing a user-friendly interface for model construction and Gibbs sampling on top of BlackJAX.
 
 In **Bamojax**, users can define a probabilistic model by specifying variables 
 as well as their associated distributions and dependencies, structured using a 
@@ -60,18 +60,17 @@ formulation, rather than the mathematical or inference details, leading to a mor
 intuitive, less error-prone, and more efficient development workflow.
 
 **Bamojax** is designed for researchers, students, and practitioners that want to
-make use of the extremely fast approximate inference offered by Blackjax, but 
+make use of the extremely fast approximate inference offered by BlackJAX, but 
 want to focus on model development instead of implementation.
 
 ## Comparison with existing tools
 
-While existing software for probabilistic modelling, such as PyMC [@pymc], can also
-interface with Blackjax for inference, this only supports using a single log-density 
-function that describes the entire probabilistic model. This precludes Gibbs sampling, 
-where individual model parameters are updated in turn, which in practice can greatly 
[…]
