### 2025-09-14 da1ccc6
Add CONTRIBUTING.md
Added instructions for external parties wanting to contribute (fixes #4 )
