+++ CONTRIBUTING.md
@@ -1,0 +1,23 @@
+# Guidelines for contributing to Bamojax
+
+Thanks you for your interest in contributing to **bamojax**! We value contributions such as:
+
+* Bug fixes
+* Documentation
+* New model implementations
+* New features that improve the pipeline anywhere from model conception to posterior inference
+
+## Reporting issues
+
+If you find a bug or have a feature request, please open a GitHub Issue and describe it clearly.
+
+## Pull requests
+If you want to make a code or documentation change, please:
+  1. Fork the repository
+  2. Create a new branch (`git checkout -b fix-bug` or `feature-new-thing`)
+  3. Make your changes
+  4. Push to your fork and open a Pull Request (PR)
+
+We will then review your PR and provide feedback or merge it with the main branch.
+
+
