### 2025-09-14 #8642 @mhinne
@kanishkan91 Great, will do!

@matt-graham thanks for your constructive pointers! I have addressed several issues today, and will continue addressing the remainder shortly (the published API and testing instructions take me a little longer to set up properly, but shouldn't take too long).
