### 2025-09-14 bf78802
Update README.md
Made consistent with paper.md
