### 2025-09-16 2312eca
Add dev options and pyproject.toml
