### 2025-09-16 #8642 @kanishkan91
Hi @alexlyttle would you be interested in reviewing this pacakge and related paper for the Journal of Open Source Software (JOSS)?

You were recommended by @warrickball
