### 2025-09-18 948ee1e
Add info on automated testing to CONTRIBUTING.md (fixes #3)
