+++ bamojax/__init__.py
@@ -1,3 +1,2 @@
-__version__ = '0.0.1'
 from . import _version
 __version__ = _version.get_versions()['version']
