### 2025-09-18 64db0b8
Add docs via mkdocs and github pages
