+++ docs/index.md
@@ -1,0 +1,6 @@
+# BaMoJAX
+
+A JAX-powered toolbox for Bayesian modelling.
+
+- **Docs**: API and examples
+- **Source**: [GitHub](https://github.com/UncertaintyInComplexSystems/bamojax)
