+++ docs/api.md
@@ -1,3 +1,11 @@
 # API reference
 
 ::: bamojax
+    handler: python
+    options:
+      show_root_heading: true
+      show_object_full_path: true
+      show_submodules: true
+      members_order: source
+      filters: ["!^_"]      # hide private members
+      inherited_members: true
