+++ mkdocs.yml
@@ -18,7 +18,7 @@ nav:
 
 plugins:
   - search
-    - gen-files:
+  - gen-files:
       scripts:
         - docs/gen_api_nav.py
   - literate-nav:
