+++ pyproject.toml
@@ -39,6 +39,14 @@ dev = [
   "pytest-cov>=5",
   "JaxKern @ git+https://github.com/UncertaintyInComplexSystems/JaxKern.git",
 ]
+docs = [
+  "mkdocs",
+  "mkdocs-material",
+  "mkdocstrings[python]",
+  "pymdown-extensions",
+  "mkdocs-gen-files",
+  "mkdocs-literate-nav"
+]
 
 [project.urls]
 Homepage = "https://github.com/UncertaintyInComplexSystems/bamojax"
