+++ .github/workflows/docs.yml
@@ -16,7 +16,13 @@ jobs:
       - uses: actions/setup-python@v5
         with:
           python-version: '3.11'
-      - run: python -m pip install -U pip
-      - run: pip install mkdocs mkdocs-material "mkdocstrings[python]" pymdown-extensions
-      - run: mkdocs build --strict
-      - run: mkdocs gh-deploy --force
+      - name: Install doc deps
+        run: |
+          python -m pip install -U pip
+          pip install -e ".[docs]"
+
+      - name: Build site
+        run: mkdocs build --strict
+
+      - name: Deploy to gh-pages
+        run: mkdocs gh-deploy --force
