+++ docs/SUMMARY.md
@@ -1,0 +1,4 @@
+# Table of contents
+
+- [Home](index.md)
+- [API](api/index.md)
