+++ mkdocs.yml
@@ -12,9 +12,6 @@ theme:
   palette:
     - scheme: default
 
-nav:
-  - Home: index.md
-  - API: api/index.md
 
 plugins:
   - search
