+++ docs/gen_api_nav.py
@@ -1,30 +1,23 @@
+# docs/gen_api_nav.py
 import mkdocs_gen_files as gen
 from pathlib import Path
 
 pkg = "bamojax"
-modules = []
+mods = []
 
 for path in sorted(Path(pkg).rglob("*.py")):
     if path.name == "__init__.py":
         continue
-    mod = ".".join(path.with_suffix("").parts)
-    modules.append(mod)
+    mods.append(".".join(path.with_suffix("").parts))
 
+# Landing page for API
 with gen.open("api/index.md", "w") as f:
     print("# API reference\n", file=f)
-    for mod in modules:
-        print(f"- [{mod}](./{mod}.md)", file=f)
+    for m in mods:
+        print(f"- [`{m}`](./{m}.md)", file=f)
 
-for mod in modules:
-    out_path = Path("api", f"{mod}.md")
-    out_path.parent.mkdir(parents=True, exist_ok=True)
-    with gen.open(out_path, "w") as f:
-        print(f"# `{mod}`\n", file=f)
-        print(f"::: {mod}\n", file=f)
-
-with gen.open("SUMMARY.md", "w") as nav:
-    print("# Table of contents\n", file=nav)
-    print("- [Home](index.md)", file=nav)
-    print("- API", file=nav)
-    for mod in modules:
-        print(f"  - [`{mod}`](api/{mod}.md)", file=nav)
+# One page per module
+for m in mods:
+    with gen.open(f"api/{m}.md", "w") as f:
+        print(f"# `{m}`\n", file=f)
+        print(f"::: {m}\n", file=f)
