+++ mkdocs.yml
@@ -25,11 +25,12 @@ plugins:
         python:
           options:
             docstring_style: google   # or "numpy"
-            show_source: true
-            source_url_pattern: "https://github.com/UncertaintyInComplexSystems/bamojax/blob/main/{path}#L{start_line}-L{end_line}"
+            show_source: true            
             members_order: source
             filters: ["!^_"]          # hide private members
             inherited_members: true
+            extra:
+              source_url_pattern: "https://github.com/UncertaintyInComplexSystems/bamojax/blob/main/{path}#L{start_line}-L{end_line}"
 
 markdown_extensions:
   - admonition
