### 2025-09-18 5bc6a24
Add link to API documentation to README.md (fixes #2)
