### 2025-09-18 04b1050
Update documentation of API
