### 2025-09-29 UncertaintyInComplexSystems/bamojax#9 @alexlyttle
Installation section of README.md could be clearer
Raising as a part of https://github.com/openjournals/joss-reviews/issues/8642

The installation section of `README.md` states that dependancies are not automatically installed, but it appears that they are.

```
pip install git+https://github.com/UncertaintyInComplexSystems/bamojax#egg=bamojax
```

Outputs,

```
...
Building wheels for collected packages: bamojax
  Building wheel for bamojax (pyproject.toml) ... done
  Created wheel for bamojax: filename=bamojax-0.2.0+30.g04b1050-py3-none-any.whl size=41756 sha256=6bdbd7364f160ca93617c25c9ffd1305223dad2db096a758a9e78d353e5ac8f4
  Stored in directory: /private/var/folders/hn/x47wb9d94z14j413yc7n8lpc0000gr/T/pip-ephem-wheel-cache-dm3wd1u0/wheels/b1/86/e1/14d9de36588201a8db45734672b6f91384bee6e8d09f996423
Successfully built bamojax
Installing collected packages: wrapt, wadler-lindig, typing-extensions, toolz, six, setuptools, opt_einsum, numpy, gast, fastprogress, decorator, cloudpickle, attrs, absl-py, scipy, ml_dtypes, jaxtyping, dm-tree, tensorflow-probability, jaxlib, jax, jaxopt, chex, optax, distrax, blackjax, bamojax
Successfully installed absl-py-2.3.1 attrs-25.3.0 bamojax-0.2.0+30.g04b1050 blackjax-1.2.4 chex-0.1.91 cloudpickle-3.1.1 decorator-5.2.1 distrax-0.1.5 dm-tree-0.1.9 fastprogress-1.0.3 gast-0.6.0 jax-0.7.2 jaxlib-0.7.2 jaxopt-0.8.5 jaxtyping-0.3.2 ml_dtypes-0.5.3 numpy-2.3.3 opt_einsum-3.4.0 optax-0.2.6 scipy-1.16.2 setuptools-80.9.0 six-1.17.0 tensorflow-probability-0.25.0 toolz-1.0.0 typing-extensions-4.15.0 wadler-lindig-0.1.7 wrapt-1.17.3
```

Could you update this to say that dependencies like `jax` and `jaxlib` are installed? If you want hardware specific versions (e.g. CUDA) then you might want to instruct the user to install these first (e.g. referring to the JAX documentation).
