### 2025-09-29 UncertaintyInComplexSystems/bamojax#10 @alexlyttle
Package does not work with JAX >= 0.7.0
Relates to https://github.com/openjournals/joss-reviews/issues/8642.

**Command:**

```python
from bamojax.base import Model
```

Raises the following exception as of JAX v0.7.0:

<summary>

```console
AttributeError: jax.interpreters.xla.pytype_aval_mappings was deprecated in JAX v0.5.0 and removed in JAX v0.7.0. jax.core.pytype_aval_mappings can be used as a replacement in most cases.
```

```console
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/***/Repos/local/bamojax-review/.venv/lib/python3.12/site-packages/bamojax/base.py", line 4, in <module>
    import distrax as dx
  File "/Users/***/Repos/local/bamojax-review/.venv/lib/python3.12/site-packages/distrax/__init__.py", line 18, in <module>
    from distrax._src.bijectors.bijector import Bijector
  File "/Users/***/Repos/local/bamojax-review/.venv/lib/python3.12/site-packages/distrax/_src/bijectors/bijector.py", line 27, in <module>
    tfb = tfp.bijectors
          ^^^^^^^^^^^^^
  File "/Users/***/Repos/local/bamojax-review/.venv/lib/python3.12/site-packages/tensorflow_probability/python/internal/lazy_loader.py", line 56, in __getattr__
    module = self._load()
             ^^^^^^^^^^^^
  File "/Users/***/Repos/local/bamojax-review/.venv/lib/python3.12/site-packages/tensorflow_probability/python/internal/lazy_loader.py", line 43, in _load
    module = importlib.import_module(self.__name__)
             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/***/miniforge3/lib/python3.12/importlib/__init__.py", line 90, in import_module
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/***/Repos/local/bamojax-review/.venv/lib/python3.12/site-packages/tensorflow_probability/substrates/jax/__init__.py", line 42, in <module>
    from tensorflow_probability.substrates.jax import bijectors
  File "/Users/***/Repos/local/bamojax-review/.venv/lib/python3.12/site-packages/tensorflow_probability/substrates/jax/bijectors/__init__.py", line 19, in <module>
    from tensorflow_probability.substrates.jax.bijectors.absolute_value import AbsoluteValue
  File "/Users/***/Repos/local/bamojax-review/.venv/lib/python3.12/site-packages/tensorflow_probability/substrates/jax/bijectors/absolute_value.py", line 17, in <module>
    from tensorflow_probability.python.internal.backend.jax.compat import v2 as tf
  File "/Users/***/Repos/local/bamojax-review/.venv/lib/python3.12/site-packages/tensorflow_probability/python/internal/backend/jax/__init__.py", line 19, in <module>
    from tensorflow_probability.python.internal.backend.jax import compat
  File "/Users/***/Repos/local/bamojax-review/.venv/lib/python3.12/site-packages/tensorflow_probability/python/internal/backend/jax/compat.py", line 17, in <module>
    from tensorflow_probability.python.internal.backend.jax import v1
  File "/Users/***/Repos/local/bamojax-review/.venv/lib/python3.12/site-packages/tensorflow_probability/python/internal/backend/jax/v1.py", line 23, in <module>
    from tensorflow_probability.python.internal.backend.jax import linalg_impl
  File "/Users/***/Repos/local/bamojax-review/.venv/lib/python3.12/site-packages/tensorflow_probability/python/internal/backend/jax/linalg_impl.py", line 23, in <module>
    from tensorflow_probability.python.internal.backend.jax import ops
  File "/Users/***/Repos/local/bamojax-review/.venv/lib/python3.12/site-packages/tensorflow_probability/python/internal/backend/jax/ops.py", line 681, in <module>
    jax.interpreters.xla.pytype_aval_mappings[onp.ndarray])
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/***/Repos/local/bamojax-review/.venv/lib/python3.12/site-packages/jax/_src/deprecations.py", line 54, in getattr
    raise AttributeError(message)
AttributeError: jax.interpreters.xla.pytype_aval_mappings was deprecated in JAX v0.5.0 and removed in JAX v0.7.0. jax.core.pytype_aval_mappings can be used as a replacement in most cases.
```

</summary>

While you note that the package is tested on JAX v0.4, that version is a year old and some users may want to use bamojax with newer versions of JAX.
