### 2025-09-29 UncertaintyInComplexSystems/bamojax#11 @alexlyttle
Fixed package dependencies
Relating to https://github.com/openjournals/joss-reviews/issues/8642.

Is there a reason for fixing `blackjax` and `distrax` versions? I understand this reduces overhead in package maintenance (e.g. keeping an eye out for braking changes). However, it is worth considering the sustainability in this decision and potential future problems.

Some dependencies in `pyproject.toml` are not fixed. This could introduce a conflict in the future if e.g. JAX introduces a breaking change which becomes incompatible with the fixed version of BlackJAX.

Fixing dependency versions (or having maximum versions) also limits how people can use your package in their own projects. For example, if BlackJAX introduced a new MCMCM kernel in a future version, one wouldn't be able to use it with bamojax without violating the requirements.

Good practice is to either fix, or provide minimum versions, for all dependencies. The latter is generally preferred for libraries you expect people to use in different contexts. The former is more appropriate if building reproducible images or applications.
