### 2025-10-02 UncertaintyInComplexSystems/bamojax#11 @mhinne
Hey @alexlyttle,

The reason for that is pragmatic; this setup works, but if newer versions of `Blackjax` are installed, this will also install newer versions of JAX itself, which breaks with `distrax`. When I started developing bamojax I did not foresee this issue.

You're right about your suggestions though, and I also agree with the related issue of #10. My plan is to remove the `distrax` dependency altogether and use a package that is more actively maintained instead, such as `numpyro`. I liked the idea of keeping the distributions and inference part separate, but to keep bamojax future-proof I think you're right and this should change.

It's not a huge change, but does require some time; I'll hope to get around to it shortly!
