### 2025-10-07 UncertaintyInComplexSystems/bamojax#11 @matt-graham
I agree with @alexlyttle's comments here - ideally the dependencies and version specifiers for a package specified in `pyproject.toml` / `setup.py` / `setup.cfg` should be the minimal set of constraints that will allow the package to run. [The compatible release `~=` version specifier](https://packaging.python.org/en/latest/specifications/version-specifiers/#compatible-release) is often useful for this purpose providing dependencies are following a semantic versioning scheme (technically JAX uses 'effort-based versioning' but specifying a compatible minor version should still broadly work).

Related to this you currently specify your dependency versions in both `setup.py` in the `install_requires` argument to `setuptools.setup` and in the `project.dependencies` key in `pyproject.toml`. I'm not actually sure how `pip` will choose to resolve when there are multiple sources of package dependencies like this, but I would say from a don't-repeat-yourself (DRY) perspective it would be better to have the dependencies specified in just one place. I would recommend `pyproject.toml` as this is the more modern standard and would make it easier to transition to other non `setuptools` build backends in future if needed.

Also the comment

https://github.com/UncertaintyInComplexSystems/bamojax/blob/04b1050001bf62e1b2f2975adcbb310112e785fb/README.md?plain=1#L26

in the README is a bit misleading as `jax` and `jaxlib` are specified as dependencies in `pyproject.toml` and `setup.py`. While flagging that users may need to manually install `jax` to allow it to access the specific device backends they are targetting is a good idea (possibly worth even linking to JAX installation documentation), I think it still worth having these included as dependencies of the package (as you do currently) as this means users will at least get something working even if it does not allow e.g. running on a GPU, when installing the package. Also functionally even if you omitted `jax` and `jaxlib` from the dependencies they would be automatically pulled in as transitive (indirect) dependencies of some your direct dependencies like `blackjax`.
