### 2025-10-07 UncertaintyInComplexSystems/bamojax#11 @mhinne
You are both absolutely right, and thank you for pointing out several tips as well. I'm actually already in the middle of migrating away from `distrax` to `numpyro` and that is going quite painlessly. I hope to have it up and running next week (it's not that much work, but I don't have that much coding time). This greatly simplifies the dependencies, and tackles several of the issues at once.

Once I am happy with these changes I'll merge them here and update the dependency management accordingly, including removing the redundant dependency listing (the `setup.py` is an artefact from my previous approach). I'll also update the README to reflect the (easier) installation procedure by then.

To be continued soon!
