### 2025-10-08 32f25b5
Replace distrax with numpyro across codebase, examples, and tests
