### 2025-10-08 940464b
Replace distrax with numpyro across codebase, examples, and tests
