### 2025-10-08 a219cc8
Replace distrax with numpyro across codebase, examples, and tests
