### 2025-10-08 d9dbdb8
Replace distrax with numpyro across codebase, examples, and tests
