### 2025-10-08 e2b4eec
Replace distrax with numpyro across codebase, examples, and tests
