### 2025-10-08 a1a66c9
Replace distrax with numpyro across codebase, examples, and tests
