### 2025-10-09 890990a
Replace distrax by numpyro in examples
