### 2025-10-09 8ae222e
Replace distrax by numpyro in examples
