### 2025-10-09 2837a6f
Replace distrax by numpyro in examples
