### 2025-10-09 9fff657
Replace distrax by numpyro in examples
