### 2025-10-09 6dc2538
Replace distrax by numpyro in examples
