### 2025-10-09 2ffe538
Replace distrax by numpyro in examples
