### 2025-10-09 0a67910
Replace distrax by numpyro in examples
