### 2025-10-09 5b59447
Replace distrax by numpyro in examples
