### 2025-10-10 4d02cc4
Update examples to use numpyro instead of distrax
