+++ README.md
@@ -4,7 +4,7 @@
 
 # Welcome to bamojax!
 
-Welcome to **bamojax**, the Bayesian modelling toolbox implemented using the Jax coding universe. **bamojax** is a probabilistic programming language (PPL), similar to Numpyro, PyMC, Stan, JAGS, and BUGS. It relies on [Blackjax](https://blackjax-devs.github.io/blackjax/) for approximate inference, on [Distrax](https://github.com/google-deepmind/distrax) for probability distributions and their essential operations. 
+Welcome to **bamojax**, the Bayesian modelling toolbox implemented using the Jax coding universe. **bamojax** is a probabilistic programming language (PPL), similar to Numpyro, PyMC, Stan, JAGS, and BUGS. It relies on [Blackjax](https://blackjax-devs.github.io/blackjax/) for approximate inference, on [Numpyro](https://num.pyro.ai/en/stable/) for probability distributions and their essential operations. 
 
 ## What sets bamojax apart?
 
@@ -36,7 +36,8 @@ import jax.random as jrnd
 from bamojax.base import Model
 from bamojax.inference import SMCInference
 from bamojax.samplers import mcmc_sampler
-import distrax as dx
+import numpyro as npr
+import numpyro.distributions as dist
 import blackjax as bjx
 
 key = jrnd.PRNGKey(0)
@@ -64,9 +65,9 @@ def link_fn(probs):
 
 #
 latent_theta = my_model.add_node('theta', 
-                                 distribution=dx.Beta(alpha=1, beta=1))
+                                 distribution=dist.Beta(concentration0=1, concentration1=1))
 observations = my_model.add_node('x', 
-                                 distribution=dx.Bernoulli, 
+                                 distribution=dist.Bernoulli, 
                                  parents=dict(probs=latent_theta), 
                                  link_fn=link_fn, 
                                  observations=x)
@@ -126,7 +127,7 @@ This effectively creates an empty DAG.
 Next, one adds nodes to the DAG one by one, or multiple at once in the case of iteration / plate notation: For instance, the following adds a latent random variable with a Gaussian distribution assigned to it, $x\sim \mathcal{N}(0,1)$:
 
 ```
-a_new_node = my_model.add_node('variable name', distribution=dx.Normal, parents=dict(loc=0.0, scale=1.0))
+a_new_node = my_model.add_node('variable name', distribution=dist.Normal, parents=dict(loc=0.0, scale=1.0))
 ```
 
 Depending on whether a user provides values for the `distribution=`, and/or `observations=` arguments in the `add_node()` method, **bamojax** derives whether the variable is:
@@ -134,9 +135,9 @@ Depending on whether a user provides values for the `distribution=`, and/or `obs
 2. Stochastic and observed,
 3. Deterministic and observed.
 
-The `parents` argument expects a dictionary, in which the keys must correspond to either the arguments of a Distrax `dx.Distribution` object or the arguments of a link function.
+The `parents` argument expects a dictionary, in which the keys must correspond to either the arguments of a Numpyro `dist.Distribution` object or the arguments of a link function.
 
-Nodes can furthermore take any Distrax bijector using the `bijector=` argument, for example to transform a real variable to a bounded domain or vice versa. In many cases, the same goal can be achieved by the link function, but sometimes a bijector is simpler to use.
+Nodes can furthermore take any Numpyro bijector using the `bijector=` argument, for example to transform a real variable to a bounded domain or vice versa. In many cases, the same goal can be achieved by the link function, but sometimes a bijector is simpler to use.
 
 ### Link functions
 
[…]
