### 2025-10-10 075875f
Update README to reflect proper dependencies (fixes #9)
