### 2025-10-10 940f4b7
Prepare for PIP install
