### 2025-10-10 370f1d8
Prepare for PIP install
