+++ pyproject.toml
@@ -10,7 +10,7 @@ build-backend = "setuptools.build_meta"
 name = "bamojax"
 dynamic = ["version"]
 description = "Bayesian Modelling with JAX"
-readme = {file = ["README.md"], content-type = "text/markdown"}
+readme = {file = "README.md", content-type = "text/markdown"}
 requires-python = ">=3.10"
 license = { file = "LICENSE" }
 authors = [
