### 2025-10-10 46ef785
Prepare for PIP install
