+++ .github/workflows/publish.yml
@@ -12,10 +12,13 @@ jobs:
     steps:
       - uses: actions/checkout@v4
         with:
+          ref: ${{ github.ref }}
           fetch-depth: 0  # Versioneer needs tags
       - uses: actions/setup-python@v5
         with:
           python-version: "3.11"
+      - name: Clean dist
+        run: -rf dist build *.egg-info
       - name: Install build deps
         run: python -m pip install --upgrade build
       - name: Build sdist+wheel
@@ -42,3 +45,5 @@ jobs:
         uses: pypa/gh-action-pypi-publish@release/v1
         with:
           repository-url: https://test.pypi.org/legacy/
+          verbose: true
+          skip-existing: true  # avoids failing on re-runs if files already there
