### 2025-10-13 c79d771
Prepare for PIP install
