### 2025-10-13 4fcd145
Add more model comparison functionality
