+++ bamojax/marginal_likelihoods.py
@@ -1,227 +1,0 @@
-
-import jax
-import jax.numpy as jnp
-import jax.random as jrnd
-
-import jaxopt
-from tqdm import tqdm
-
-from jax.scipy.special import logsumexp
-from jax.tree_util import tree_flatten, tree_unflatten, tree_map
-
-
-import numpyro as npr
-import numpyro.distributions as dist
-from numpyro.distributions import Distribution, TransformedDistribution
-import numpyro.distributions.transforms as nprb
-
-from jaxtyping import Float
-from typing import Callable, Dict
-
-from bamojax.base import Model
-
-# TODO: implement testing for these functions
-
-def iid_likelihood(L: Callable):
-    r"""
-    
-    We typically have multiple observations and assume the likelihood factorizes 
-    as: 
-
-    $$    
-        \log p\left(Y \mid \theta\right) = \sum_{i=1}^N \log p\left(y_i \mid \theta\right) \enspace.
-    $$
-
-    """
-    return lambda x: jnp.sum(L()(x))
-
-#
-def naive_monte_carlo(key, 
-                      model: Model, 
-                      num_prior_draws: int = 1_000, 
-                      num_chunks: int = 5,
-                      iid_obs: bool = True,
-                      pb = True) -> Float:   
-    r"""The Naive Monte Carlo (NMC) estimator
-
-    The marginal likelihood is defined as 
-    $$
-        p(D) = \int_\Theta p\left(D \mid \theta\right) p(\theta) \,\text{d}\theta \enspace .
-    $$
-
-    In NMC we draw samples from the prior and approximate the ML as
-    $$
-        p(D) \approx \frac{1}{N} \sum_{i=1}^N p\left(D \mid \theta_i\right) \enspace,
-    $$  with $\theta_i \sim p(\theta)$.
-
-    In nontrivial models, we need a *large* $N$ for this approximation to be 
-    reasonable.
-
-    """
-
-    if iid_obs:
-        loglikelihood_fn = iid_likelihood(model.loglikelihood_fn)
-    else:
-        loglikelihood_fn = model.loglikelihood_fn
-
-    loglikelihoods = jnp.zeros((num_prior_draws, 
-                                num_chunks))
-    
-    # We don't want to vmap this loop, as the reason for the loop is to avoid
-    # running out of memory!
-    for i in tqdm(range(num_chunks), disable=not pb):
-        key, subkey = jrnd.split(key)
-        keys = jrnd.split(subkey, num_prior_draws)
-        prior_draws = jax.vmap(model.sample_prior)(keys)
-        loglikelihoods = loglikelihoods.at[:, i].set(jax.vmap(loglikelihood_fn)(prior_draws))
-    return logsumexp(loglikelihoods.flatten()) - jnp.log(num_prior_draws*num_chunks)
-
-#
-def harmonic_mean(key, model: Model) -> Float:
-    raise NotImplementedError
-
-def generalized_harmonic_mean(key, model: Model) -> Float:
-    raise NotImplementedError
-
-#
-def importance_sampling(key, 
-                        model: Model, 
-                        g_IS: Distribution,
-                        num_samples: int = 1_000,
-                        iid_obs: bool = True) -> Float:
-    
-    r"""Importance sampling routine for a given BayesianModel.
-
-    Importance sampling is based around the following approximation to the log
-    marginal likelihood [Gronau et al., 2017]:
-
-    $$
-    p(D) \approx \frac{1}{N} \sum_{i=1}^N p\left(D \mid \theta_i\right) \frac{p(\theta_i)}{g_IS(\theta_i)}\enspace,
-    $$
-    with $\theta_i \sim g_IS(\theta)$
-
[…]
