### 2025-10-13 eb319e6
Add more model comparison functionality
