### 2025-10-14 ca3d35b
Update inference.py
