+++ bamojax/__init__.py
@@ -1,5 +1,2 @@
 from . import _version
 __version__ = _version.get_versions()['version']
-
-from . import _version
-__version__ = _version.get_versions()['version']
