### 2025-10-14 a28e78c
Add setup.py
