### 2025-10-14 e6ed66e
Update README to include model comparison features
