### 2025-10-15 5a6585a
Add THAMES estimator
