### 2025-10-15 f6f1d99
Update bridge vs THAMES vs SMC example
