### 2025-10-16 #8642 @matt-graham
Hi @mhinne and @kanishkan91. Thanks @mhinne for the updates to relax the JAX dependency requirement / switching to NumPyro - I've had a look through updated package and tested installing and running quick start example and everything works great for me. I have now completed review checklist and from my side every looks good to accept 😄
