### 2025-10-23 9bec07d
Add marginal likelihood estimators to THAMES example
