+++ paper/paper.bib
@@ -57,6 +57,22 @@
   url = 	 {https://proceedings.mlr.press/v9/murray10a.html}
 }
 
+@article{numpyro,
+  title={Composable Effects for Flexible and Accelerated Probabilistic Programming in NumPyro},
+  author={Phan, Du and Pradhan, Neeraj and Jankowiak, Martin},
+  journal={arXiv preprint arXiv:1912.11554},
+  year={2019}
+}
+
+@software{oryx,
+  author       = {Oryx contributors (jax-ml)},
+  title        = {Oryx: Probabilistic programming and deep learning in JAX},
+  year         = {2025},
+  version      = {0.2.9},
+  url          = {https://github.com/jax-ml/oryx},
+  note         = {[Software]}
+}
+
 
 @article{pymc,
   title = {{PyMC}: A Modern and Comprehensive Probabilistic Programming Framework in {P}ython},
