### 2025-10-27 257b988
Update CITATION.cff
