+++ CITATION.cff
@@ -17,4 +17,4 @@ keywords:
   - probabilistic programming
   - JAX
   - Blackjax
-  - Python
+  - Bayesian model comparison
