### 2025-10-27 4e0a9e5
Update small mistakes in paper
