+++ paper/paper.bib
@@ -65,12 +65,11 @@
 }
 
 @software{oryx,
-  author       = {Oryx contributors (jax-ml)},
-  title        = {Oryx: Probabilistic programming and deep learning in JAX},
-  year         = {2025},
-  version      = {0.2.9},
-  url          = {https://github.com/jax-ml/oryx},
-  note         = {[Software]}
+  author = {{The Oryx Authors}},
+  title = {Oryx: Probabilistic Programming and Deep Learning in JAX},
+  howpublished = {\url{https://github.com/jax-ml/oryx}},
+  url = {https://github.com/jax-ml/oryx},
+  year = {2022}
 }
 
 
