### 2025-10-27 ba469ff
Add JOSS citation info
