### 2025-10-27 2f15eb8
Add JOSS citation info
