### 2025-10-27 4d772b1
Improve GP factory broadcasting
