### 2025-10-27 ff860ef
Fix PyMC reference
