+++ paper/paper.bib
@@ -75,7 +75,7 @@
 
 @article{pymc,
   title = {{PyMC}: A Modern and Comprehensive Probabilistic Programming Framework in {P}ython},
-  author = {Abril-Pla O, Andreani V, Carroll C, Dong L, Fonnesbeck CJ, Kochurov M, Kumar R, Lao J, Luhmann CC, Martin OA, Osthege M, Vieira R, Wiecki T, Zinkov R.},
+  author = {Oriol Abril-Pla and Virgile Andreani and Colin Carroll and Larry Dong and Christopher J. Fonnesbeck and Maxim Kochurov and Ravin Kumar and Junpeng Lao and Christian C. Luhmann and Osvaldo A. Martin and Michael Osthege and Ricardo Vieira and Thomas Wiecki and Robert Zinkov},
   journal = {{PeerJ} Computer Science},
   volume = {9},
   number = {e1516},
