### 2025-11-04 6d3dbb1
Update the dict flattening to support singleton vector variables
